pyproject.toml
[tool.poetry]
name = "drf-api"
version = "0.1.0"
description = ""
authors = ["Ryan Pilon <ryanpilon1995@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.4"
djangorestframework = "^3.12.2"
djangorestframework-simplejwt = "^4.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
